# vitis-vision-front-end/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vitis Vision Service",
    "description": "OpenAPI documentation for Vitis Vision Service",
    "contact": {
      "name": "Vitis Vision",
      "url": "https://vitisvision.com",
      "email": "vitisvision@gmail.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8181",
      "description": "Local Server"
    }
  ],
  "security": [
    {
      "Bearer Auth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/vineyards/{vineyardId}/blocks/{blockId}": {
      "get": {
        "tags": [
          "Block"
        ],
        "summary": "Get a block in a vineyard",
        "description": "Get a block in a vineyard with the provided details",
        "operationId": "getBlock",
        "parameters": [
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBlockResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Block"
        ],
        "summary": "Update a block in a vineyard",
        "description": "Update a block in a vineyard with the provided details",
        "operationId": "updateBlock",
        "parameters": [
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBlockResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Block"
        ],
        "summary": "Delete a block in a vineyard",
        "description": "Delete a block in a vineyard with the provided details",
        "operationId": "deleteBlock",
        "parameters": [
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vineyards/{vineyardId}/blocks": {
      "get": {
        "tags": [
          "Block"
        ],
        "summary": "Get all blocks in a vineyard",
        "description": "Get all blocks in a vineyard with the provided details",
        "operationId": "getBlocks",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageableResponseListBlockResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Block"
        ],
        "summary": "Create a new block in a vineyard",
        "description": "Create a new block in a vineyard with the provided details",
        "operationId": "creteBlock",
        "parameters": [
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBlockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vineyards/{vineyardId}/blocks/{blockId}/_report": {
      "post": {
        "tags": [
          "Block"
        ],
        "summary": "Generate a report for a block in a vineyard",
        "description": "Generate a report for a block in a vineyard with the provided details",
        "operationId": "generateBlockReport",
        "parameters": [
          {
            "name": "vineyardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 EARLY_HINTS",
              "103 CHECKPOINT",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "302 MOVED_TEMPORARILY",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "305 USE_PROXY",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 PAYLOAD_TOO_LARGE",
              "413 REQUEST_ENTITY_TOO_LARGE",
              "414 URI_TOO_LONG",
              "414 REQUEST_URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "419 INSUFFICIENT_SPACE_ON_RESOURCE",
              "420 METHOD_FAILURE",
              "421 DESTINATION_LOCKED",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "BlockRequest": {
        "required": [
          "name",
          "partitioningType",
          "rowOrientation",
          "rowSpacing",
          "trellisSystemType",
          "vineSpacing"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "partitioningType": {
            "type": "string"
          },
          "rowSpacing": {
            "maximum": 1000,
            "exclusiveMaximum": false,
            "minimum": 0,
            "exclusiveMinimum": false,
            "type": "number",
            "format": "double"
          },
          "vineSpacing": {
            "maximum": 1000,
            "exclusiveMaximum": false,
            "minimum": 0,
            "exclusiveMinimum": false,
            "type": "number",
            "format": "double"
          },
          "rowOrientation": {
            "type": "string"
          },
          "trellisSystemType": {
            "type": "string"
          }
        }
      },
      "ApiResponseBlockResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "$ref": "#/components/schemas/BlockResponse"
          }
        }
      },
      "BlockResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "partitioningType": {
            "type": "string"
          },
          "rowSpacing": {
            "type": "number",
            "format": "double"
          },
          "vineSpacing": {
            "type": "number",
            "format": "double"
          },
          "rowOrientation": {
            "type": "string"
          },
          "trellisSystemType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUpdatedAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "lastUpdatedBy": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "BlockReportRequest": {
        "required": [
          "aggregationInterval"
        ],
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "aggregationInterval": {
            "maximum": 1440,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResponsePageableResponseListBlockResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "$ref": "#/components/schemas/PageableResponseListBlockResponse"
          }
        }
      },
      "PageableResponseListBlockResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockResponse"
            }
          },
          "pageMetadata": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Auth": {
        "type": "http",
        "description": "JWT bearer token security scheme",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}